requirements.txt
rich>=10.0.0
questionary>=1.10.0
pyttsx3>=2.90
python-dateutil>=2.8.2
prompt_toolkit>=3.0.0 